+++ src/aks-preview/azext_aks_preview/_consts.py
@@ -42,6 +42,7 @@ CONST_VIRTUAL_MACHINES = "VirtualMachines"
 
 # vm size
 CONST_DEFAULT_NODE_VM_SIZE = "Standard_DS2_v2"
+CONST_DEFAULT_AUTOMATIC_SKU_NODE_VM_SIZE = "Standard_DS4_v2"
 CONST_DEFAULT_WINDOWS_NODE_VM_SIZE = "Standard_D2s_v3"
 
 # workload runtime
@@ -66,6 +67,10 @@ CONST_LOAD_BALANCER_SKU_STANDARD = "standard"
 CONST_MANAGED_CLUSTER_SKU_TIER_FREE = "free"
 CONST_MANAGED_CLUSTER_SKU_TIER_STANDARD = "standard"
 CONST_MANAGED_CLUSTER_SKU_TIER_PREMIUM = "premium"
+
+# ManagedClusterSKU Name
+CONST_MANAGED_CLUSTER_SKU_NAME_BASE = "base"
+CONST_MANAGED_CLUSTER_SKU_NAME_AUTOMATIC = "automatic"
 
 CONST_OUTBOUND_MIGRATION_MULTIZONE_TO_NATGATEWAY_MSG = (
     "Warning: this AKS cluster has multi-zonal nodepools, but NAT Gateway is not currently zone redundant. "